LayoutTests/platform/ios-simulator/editing/shadow/shadow-selection-not-exported-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
This test ensures shadow root won't expose to JS layer.

BEFORE
CONTAINER
AFTER
